+++ mobile/src/main/java/com/mercandalli/android/home/train/TrainTrafficCardView.kt
@@ -68,7 +68,7 @@ class TrainTrafficCardView @JvmOverloads constructor(
 
     private fun createTrainTrafficListener(): TrainManager.TrainTrafficListener {
         return object : TrainManager.TrainTrafficListener {
-            override fun onTrainTrafficChanged(trainTrafficType: Long) {
+            override fun onTrainTrafficChanged(trainTrafficType: Int) {
                 sync()
             }
         }
